Add ability to run a custom hook during k8s test setup (#38)
This change adds the ability to specific a custom hook that will run after the k8s setup is completed but before tests are run. This hook can be used for such things as running pre-test suite setup that is not specific to a single test, or for batching the setup for multiple tests for efficiency.