crds: Add support for deploying CRDs before other resources.
This handles an inherent race condition in kubernetes in which `kubectl apply` applies manifests too fast such that if you apply a new CRD and a resource that uses that same CRD in the same
`apply` call, the application will fail with `no matches for kind <foo> in version <foo version>`. The traditional fix is to apply CRDs first, and then wait for them to be established before applying the rest of the resources.

This is accomplished by adding a new binary called `sorter` into the fold. `sorter` will sort the yamls based on things that should be applied first (ie, namespaces should always be applied first, etc) and also can optionally filter based on kind. If we specify waiting for CRDs, we will use apply CRDs first using the sorter, wait, then apply the rest of the manifests.

Of note, kustomize offers basic sorting, but it doesn't handle the majority of resource types and being beholden to kustomize's sorting means we cannot alter it. So we disable kustomize sorting in favor of our own.

Also of note, when deleting manifests, we must delete resources in the opposite order because if we disable a CRD first, it will delete all usages of the CRD causing delete to fail because it will try to delete a non-existing resource.